Refactor agent chat panel helpers into dedicated modules
