+++ app/ui/agent_chat_panel/__init__.py
@@ -1,7 +1,8 @@
 """Agent chat panel package."""
 
+from .confirm_preferences import RequirementConfirmPreference
 from .execution import AgentCommandExecutor, ThreadedAgentCommandExecutor
-from .panel import AgentChatPanel, RequirementConfirmPreference
+from .panel import AgentChatPanel
 from .paths import history_path_for_documents, settings_path_for_documents
 from .project_settings import AgentProjectSettings
